consolidate frontend build files
