+++ frontend_build/src/webpack.config.js
@@ -9,8 +9,8 @@
 
 var path = require('path');
 var webpack = require('webpack');
-var logging = require('../assets/src/logging');
-var readBundlePlugins = require('../assets/src/read_bundle_plugins');
+var logging = require('./logging');
+var readBundlePlugins = require('./read_bundle_plugins');
 
 
 // Mappings for libraries that we bundle in the Kolibri core app.
@@ -27,6 +27,6 @@ var libs = function(kolibri_name) {
   };
 };
 
-var bundles = readBundlePlugins(path.dirname(__dirname), libs);
+var bundles = readBundlePlugins(path.join(__dirname, '..', '..'), libs);
 
 module.exports = bundles;
