+++ package.json
@@ -4,14 +4,14 @@
   "description": "Development utilities for Kolibri",
   "scripts": {
     "test": "npm run test-mocha && npm run test-karma",
-    "test-karma": "node ./node_modules/karma/bin/karma start karma_config/karma.conf.js --single-run",
-    "test-mocha": "node ./node_modules/mocha/bin/mocha assets/test/*.js",
+    "test-karma": "node ./node_modules/karma/bin/karma start ./karma_config/karma.conf.js --single-run",
+    "test-mocha": "node ./node_modules/mocha/bin/mocha ./frontend_build/test/*.js",
     "coverage": "npm run test-mocha-cov && npm run test-karma-cov",
-    "test-mocha-cov": "node ./node_modules/istanbul/lib/cli.js cover ./node_modules/mocha/bin/_mocha assets/test/*.js --report lcovonly -- -R spec",
-    "test-karma-cov": "./node_modules/karma/bin/karma start karma_config/karma.conf.ci.js",
-    "build": "webpack --config webpack_config/webpack.config.prod.js",
-    "watch": "webpack --config webpack_config/webpack.config.dev.js --watch --progress",
-    "test-karma:watch": "./node_modules/karma/bin/karma start karma_config/karma.conf.js"
+    "test-mocha-cov": "node ./node_modules/istanbul/lib/cli.js cover ./node_modules/mocha/bin/_mocha frontend_build/test/*.js --report lcovonly -- -R spec",
+    "test-karma-cov": "./node_modules/karma/bin/karma start ./karma_config/karma.conf.ci.js",
+    "build": "webpack --config ./frontend_build/src/webpack.config.prod.js",
+    "watch": "webpack --config ./frontend_build/src/webpack.config.dev.js --watch --progress",
+    "test-karma:watch": "./node_modules/karma/bin/karma start ./karma_config/karma.conf.js"
   },
   "repository": {
     "type": "git",
